requirements.txt
# 프로젝트 인원 추천 챗봇 의존성 패키지
# ==========================================

# 웹 인터페이스
streamlit>=1.28.0

# LangChain 생태계 (AI 에이전트)
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.2.0
faiss-cpu>=1.7.4
langgraph>=0.0.20

# 데이터 처리 및 분석
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# 자연어 처리 (한국어 지원)
sentence-transformers>=2.2.0

# 데이터 시각화
plotly>=5.15.0
matplotlib>=3.7.0
wordcloud>=1.9.0

# 환경 변수 관리
python-dotenv>=1.0.0

# 추가 유틸리티
tqdm>=4.65.0  # 진행률 표시
